Move persistence into own file.
